% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd_fo.R
\name{data_process_fo}
\alias{data_process_fo}
\title{Process first occurrence (FO) data in UKB ----}
\usage{
data_process_fo(data, fo_date_col, censored = "2025-01-01")
}
\arguments{
\item{data}{Dataframe with UKB baseline / FO / death dates.}

\item{fo_date_col}{Column name (string) of first occurrence date to use.}

\item{censored}{Censoring date (character, "YYYY-MM-DD").}
}
\value{
Dataframe with eid, {disease_prefix}_status, {disease_prefix}_time.
}
\description{
Process first occurrence (FO) data in UKB ----
}
