% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dx.R
\name{dx_login}
\alias{dx_login}
\title{Log in to DNAnexus}
\usage{
dx_login(token = NULL)
}
\arguments{
\item{token}{Optional. A DNAnexus authentication token.}
}
\value{
None (runs system command).
}
\description{
This is a wrapper for \code{dx login}. It supports both interactive login
(if no token provided) and token-based login.
}
