% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{format_p_value}
\alias{format_p_value}
\title{Helper function to format p-values according to common publication standards}
\usage{
format_p_value(p_value)
}
\arguments{
\item{p_value}{Numeric p-value}
}
\value{
Formatted p-value string
}
\description{
Helper function to format p-values according to common publication standards
}
