% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_1.R
\name{leo.table1.help}
\alias{leo.table1.help}
\title{Table 1 Help}
\usage{
leo.table1.help()
}
\description{
This function provides help on how to use the Table 1 functions, including guidance on how to create Table 1 and select appropriate normality tests.
}
