% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_1.R
\name{leo.table1.save}
\alias{leo.table1.save}
\title{Save Table 1 to Word or CSV}
\usage{
leo.table1.save(table, filename)
}
\arguments{
\item{table}{A table object (e.g., from CreateTableOne).}

\item{filename}{A character string indicating the output file name (e.g., "table1.csv" or "table1.docx").}
}
\description{
This function saves the Table 1 output as a Word or CSV file, depending on the file extension.
}
