% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_1.R
\name{leo.table1.step1}
\alias{leo.table1.step1}
\title{Step 1: Check Normality for Variables}
\usage{
leo.table1.step1(df, id_to_exclude = c("eid", "iri_time"), num_var = NULL)
}
\arguments{
\item{df}{A data frame or tibble containing the variables to be tested.}

\item{num_var}{user-provided numeric variables to test for normality. Default is NULL.}

\item{id}{A character string specifying the column name for patient ID. Default is "eid".}
}
\value{
A data frame with variables, p-values for normality test, and normality status.
}
\description{
This function checks the normality of each variable in the dataset and returns a summary.
}
\examples{
# leo.table1.step1(t1d_cohort, id_to_exclude = c("eid", "iri_time"), num_var = c("age", "tdi", "score_diet", "TyG"))
}
