% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_1.R
\name{leo.table1.step2}
\alias{leo.table1.step2}
\title{Step 2: Generate Table 1 for Cohort Study}
\usage{
leo.table1.step2(
  df,
  var_all,
  var_cat,
  var_non,
  strata = NULL,
  var_exact = NULL,
  compare_test = F,
  includeNA = F,
  showAllLevels = F,
  verbose = T
)
}
\arguments{
\item{df}{A data frame or tibble containing the clinical variables.}

\item{var_all}{Vector. All var to be included in table 1}

\item{var_cat}{Vector. Categorical variables.}

\item{var_non}{Vector. Non-normal continuous variables. (Process in \code{print})}

\item{strata}{Character, the column name to stratify by. Use "none" to disable stratification.}

\item{var_exact}{Vector. var_cat that needs fisher exact test. (Process in \code{print})}

\item{compare_test}{Logical, if TRUE, performs group comparison (only works when \code{strata} is not "none").}

\item{includeNA}{Logical, if TRUE, includes NAs as a level in categorical variables.}

\item{showAllLevels}{Logical, if TRUE, all levels of factor variables are shown.}

\item{verbose}{Logical, if TRUE, prints the table.}
}
\value{
Printed table1
}
\description{
This function generates a descriptive table for clinical cohort studies, with options to stratify by group and perform group comparisons.
}
