% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{leo_age}
\alias{leo_age}
\title{Recode age (years)
https://biobank.ndph.ox.ac.uk/showcase/field.cgi?id=21003}
\usage{
leo_age(age, ...)
}
\arguments{
\item{age}{Numeric/character vector for age (years) from p21003.}
}
\value{
Numeric vector of age in years.
}
\description{
Recode age (years)
https://biobank.ndph.ox.ac.uk/showcase/field.cgi?id=21003
}
