% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{leo_BMI}
\alias{leo_BMI}
\title{Recode BMI into categories (Underweight/Normal weight/Overweight/Obese)
https://biobank.ndph.ox.ac.uk/showcase/field.cgi?id=21001}
\usage{
leo_BMI(BMI, ...)
}
\arguments{
\item{BMI}{Numeric/character vector of BMI (kg/m^2) from p21001.}
}
\value{
Factor with levels "Underweight", "Normal weight", "Overweight", "Obese";
NA if invalid or missing.
}
\description{
Recode BMI into categories (Underweight/Normal weight/Overweight/Obese)
https://biobank.ndph.ox.ac.uk/showcase/field.cgi?id=21001
}
