% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{leo_career}
\alias{leo_career}
\title{Recode employment status}
\usage{
leo_career(p6142_i0, ...)
}
\arguments{
\item{p6142_i0}{Current employment status (UKB 6142, instance 0).
Values:
1 = In paid employment or self-employed
2 = Retired
3 = Looking after home and/or family
4 = Unable to work because of sickness or disability
5 = Unemployed
6 = Doing unpaid or voluntary work
7 = Full or part-time student
-7 = None of the above
-3 = Prefer not to answer}

\item{...}{Reserved for future use.}
}
\value{
Factor with categories "Paid employment" and "Not paid employment"; NA for special codes.
}
\description{
Recode employment status into paid vs not-paid employment
https://biobank.ndph.ox.ac.uk/showcase/field.cgi?id=6142
}
