% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_plot.r
\name{leo_dag_write.leo_dag}
\alias{leo_dag_write.leo_dag}
\title{Write a \code{leo_dag} object to disk}
\usage{
\method{leo_dag_write}{leo_dag}(x, figure_dir, output_dir, width = 8, height = 6, ...)
}
\arguments{
\item{x}{A \code{leo_dag} object.}

\item{figure_dir}{Figure output directory. Must be provided explicitly.}

\item{output_dir}{Table output directory. Must be provided explicitly.}

\item{width}{Figure width.}

\item{height}{Figure height.}

\item{...}{Reserved for S3 methods.}
}
\value{
Invisibly returns output file paths.
}
\description{
Write a \code{leo_dag} object to disk
}
