% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dag_plot.r
\name{leo_dag_write}
\alias{leo_dag_write}
\title{Write outputs for a \code{leo_dag} object}
\usage{
leo_dag_write(x, ...)
}
\arguments{
\item{x}{A \code{leo_dag} object.}

\item{...}{Reserved for S3 methods.}
}
\description{
Write outputs for a \code{leo_dag} object
}
