% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{leo_drinking_consumption}
\alias{leo_drinking_consumption}
\title{Recode drinking consumption}
\usage{
leo_drinking_consumption(...)
}
\value{
Integer vector (0/1/2; NA for negative/special codes or others).
}
\description{
This is more sophisticated indictor shoing if one is exccessively drink,
cobining thinking: red_wine white wine, beer, spirits, and fortified wine Numeric/character vector, usually from p20117.
}
