% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{leo_education}
\alias{leo_education}
\title{Recode UKB qualifications into 3 categories}
\usage{
leo_education(p6138_i0, ...)
}
\arguments{
\item{p6138_i0}{UKB 6138 qualifications string (https://biobank.ndph.ox.ac.uk/ukb/field.cgi?id=6138).}

\item{...}{Reserved for future use.}
}
\value{
Character vector with values: "Degree", "No degree", or "Unknown".
}
\description{
Recode UKB field 6138 (Qualifications; multiple selections separated by "|")
into Degree / No degree / Unknown.
}
