% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{leo_eGDR}
\alias{leo_eGDR}
\title{Calculate Estimated Glucose Disposal Rate (eGDR)}
\usage{
leo_eGDR(hba1c, waist, hypertension, ...)
}
\arguments{
\item{hba1c}{A numeric vector of HbA1c values (in mmol/mol).}

\item{waist}{A numeric vector of waist circumferences (in cm).}

\item{hypertension}{A numeric vector of hypertension status (1 for yes, 0 for no).}
}
\value{
A numeric vector of eGDR values.
}
\description{
This function calculates the eGDR using waist circumference, hypertension status, and HbA1c values.
It works element-wise for vectors of HbA1c, waist, and hypertension.
}
\note{
eGDR: estimated glucose disposal rate
\itemize{
\item eGDR = 21.158 + (-0.09 * waist circumference \link{cm}) + (−3.407 * hypertension \link{yes=1/no=0}) + (−0.551 * HbA1c (\%))
-- Ref1: https://diabetesjournals.org/care/article/36/8/2280/32950/Use-of-the-Estimated-Glucose-Disposal-Rate-as-a (Diabetes Care, 2013 July)
-- Ref2: https://pmc.ncbi.nlm.nih.gov/articles/PMC11439291/#Sec2 (Cardiovasc Diabetol, 2024 Sep)
In UKB, the HbA1c is measured in mmol/mol (IFCC), so we need to convert it to \% (NGSP)
\item HbA1c (mmol/mol) (Field ID: 30750)
-- NGSP (\%) = \link{0.09148 * IFCC} + 2.152 (Ref: https://ngsp.org/ifcc.asp)
\item waist circumference (Field ID: 48)
}
}
\references{
\enumerate{
\item \url{https://diabetesjournals.org/care/article/36/8/2280/32950/Use-of-the-Estimated-Glucose-Disposal-Rate-as-a}
\item \url{https://pmc.ncbi.nlm.nih.gov/articles/PMC11439291/#Sec2}
}
}
