% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{leo_ethnicity_finer}
\alias{leo_ethnicity_finer}
\title{Recode ethnicity into finer categories (top-level 1–6)
https://biobank.ndph.ox.ac.uk/showcase/field.cgi?id=21000}
\usage{
leo_ethnicity_finer(ethnicity, ...)
}
\arguments{
\item{ethnicity}{Numeric/character vector, usually from p21000.}
}
\value{
Factor with levels:
"White", "Mixed", "Asian or Asian British", "Black or Black British",
"Chinese", "Other ethnic group"; NA for -1/-3/missing/other unexpected codes.
}
\description{
Recode ethnicity into finer categories (top-level 1–6)
https://biobank.ndph.ox.ac.uk/showcase/field.cgi?id=21000
}
