% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{leo_gender}
\alias{leo_gender}
\title{Recode sex (Male/Female)
https://biobank.ndph.ox.ac.uk/showcase/field.cgi?id=31}
\usage{
leo_gender(gender, ...)
}
\arguments{
\item{gender}{Numeric/character vector, usually from p31.}
}
\value{
Factor with levels "Female", "Male"; NA for negative/special codes.
}
\description{
Recode sex (Male/Female)
https://biobank.ndph.ox.ac.uk/showcase/field.cgi?id=31
}
