% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{leo_hba1c}
\alias{leo_hba1c}
\title{Recode HbA1c (IFCC, mmol/mol)
https://biobank.ndph.ox.ac.uk/showcase/field.cgi?id=30750}
\usage{
leo_hba1c(hba1c, ...)
}
\arguments{
\item{hba1c}{Numeric/character vector of HbA1c values (mmol/mol, IFCC), usually from p30750.}

\item{...}{Reserved for future use.}
}
\value{
Numeric vector of HbA1c in mmol/mol (IFCC), with special codes as NA.
}
\description{
UK Biobank HbA1c is stored in mmol/mol (IFCC). This function performs:
\itemize{
\item numeric coercion,
\item recoding UKB special codes (-3/-1) to NA,
\item dropping non-positive values to NA (conservative guard).
}
}
