% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{leo_hba1c_percent}
\alias{leo_hba1c_percent}
\title{Convert HbA1c from IFCC (mmol/mol) to NGSP (\%)
https://ngsp.org/ifcc.asp}
\usage{
leo_hba1c_percent(hba1c, ...)
}
\arguments{
\item{hba1c}{Numeric/character vector of HbA1c values (mmol/mol, IFCC), usually from p30750.}

\item{...}{Reserved for future use.}
}
\value{
Numeric vector of HbA1c in \% (NGSP), NA preserved.
}
\description{
Conversion formula:
\deqn{NGSP(\%) = 0.09148 \times IFCC(mmol/mol) + 2.152}
}
