% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{leo_household_income_2}
\alias{leo_household_income_2}
\title{Recode household income (UKB 738) into readable categories (2 categories)}
\usage{
leo_household_income_2(p738_i0, ...)
}
\arguments{
\item{p738_i0}{Household income before tax (UKB 738, instance 0).}

\item{...}{Reserved for future use.}
}
\value{
Character vector
}
\description{
The five levels of the average total before-tax household income were classified
as “less than” or “equal to or above” £31,000, which is closest to the UK median
household income in October 2009 (£27,530)
\itemize{
\item Reference: \emph{Air Pollution, Genetic Factors, and the Risk of Lung Cancer A Prospective Study in the UK Biobank}
}
}
