% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{leo_new_clinical_indicators}
\alias{leo_new_clinical_indicators}
\title{Generate New Clinical Indicators}
\usage{
leo_new_clinical_indicators(
  df,
  types = c("eGDR", "TyG"),
  type_id = FALSE,
  remove_assist = TRUE,
  ...
)
}
\arguments{
\item{df}{A dataframe containing the required columns for calculation.}

\item{types}{A vector of indicator types, options are "eGDR" or "TyG".}

\item{remove_assist}{Logical, whether to remove the assisting columns from the dataframe (default TRUE).}

\item{...}{Additional parameters passed to the specific indicator calculation functions.}
}
\value{
A dataframe containing the new clinical indicators.
}
\description{
This function applies specific functions to generate new clinical indicators.
It is designed to work with vectors (e.g., columns of a dataframe).
}
\examples{
df <- data.frame(
  hba1c = c(48, 55, 60),
  waist = c(80, 85, 90),
  hypertension = c(1, 0, 1),
  triglycerides = c(1.5, 1.8, 2.0),
  glucose = c(5.0, 5.5, 6.0)
)
leo_new_clinical_indicators(df, types = c("eGDR", "TyG"))
}
