% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{leo_tdi}
\alias{leo_tdi}
\title{Recode Townsend deprivation index (TDI)
https://biobank.ndph.ox.ac.uk/showcase/field.cgi?id=22189}
\usage{
leo_tdi(tdi, ...)
}
\arguments{
\item{tdi}{Numeric/character vector, usually from p22189.}
}
\value{
Numeric vector, with UKB special codes -3/-1 recoded to NA.
}
\description{
Recode Townsend deprivation index (TDI)
https://biobank.ndph.ox.ac.uk/showcase/field.cgi?id=22189
}
