% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{leo_TyG}
\alias{leo_TyG}
\title{Calculate Triglyceride-Glucose Index (TyG)}
\usage{
leo_TyG(triglycerides, glucose, ...)
}
\arguments{
\item{triglycerides}{A numeric vector of triglyceride values (in mmol/L).}

\item{glucose}{A numeric vector of glucose values (in mmol/L).}
}
\value{
A numeric vector of TyG indices.
}
\description{
This function calculates the TyG index using triglyceride and glucose values.
It works element-wise for vectors of triglycerides and glucose.
}
\note{
TyG: triglyceride-glucose index
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3662289/}
}
