% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{recode_na31}
\alias{recode_na31}
\title{Helper function to recode UKB special values (-3/-1) to NA}
\usage{
recode_na31(x)
}
\description{
Helper function to recode UKB special values (-3/-1) to NA
}
\details{
Used for UK Biobank data-codings where -1 = "Do not know" and -3 = "Prefer not to answer".
}
\keyword{internal}
