% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{recode_na31neg10}
\alias{recode_na31neg10}
\title{Helper function to recode UKB special values (-3/-1 to NA and -10 to 0)}
\usage{
recode_na31neg10(x)
}
\description{
Helper function to recode UKB special values (-3/-1 to NA and -10 to 0)
}
\details{
For UK Biobank fields where -10 encodes "Less than once" and should be treated as 0;
-1 = "Do not know", -3 = "Prefer not to answer".
}
\keyword{internal}
