% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leo.ukb-package.R
\docType{package}
\name{leo.ukb-package}
\alias{leo.ukb}
\alias{leo.ukb-package}
\title{leo.ukb: LEO package for UK Biobank database analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package is used to analyze UK Biobank data. But it may be extended to other clincial studies.
}
\author{
\strong{Maintainer}: Ao Lu \email{luao@stu.cqmu.edu.cn} (\href{https://orcid.org/0009-0001-0927-4468}{ORCID})

}
\keyword{internal}
