% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{leo_drinking_status}
\alias{leo_drinking_status}
\title{Recode drinking status (Never/Previous/Current)
https://biobank.ndph.ox.ac.uk/showcase/field.cgi?id=20117}
\usage{
leo_drinking_status(drinking_status, ...)
}
\arguments{
\item{drinking_status}{Numeric/character vector, usually from p20117.}
}
\value{
Factor with levels "Never", "Previous", "Current";
NA for negative/special codes or others.
}
\description{
Recode drinking status (Never/Previous/Current)
https://biobank.ndph.ox.ac.uk/showcase/field.cgi?id=20117
}
