% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{leo_eGDR}
\alias{leo_eGDR}
\title{Calculate Estimated Glucose Disposal Rate (eGDR)}
\usage{
leo_eGDR(hba1c, waist, hypertension, ...)
}
\arguments{
\item{hba1c}{A numeric vector of HbA1c values (in mmol/mol).}

\item{waist}{A numeric vector of waist circumferences (in cm).}

\item{hypertension}{A numeric vector of hypertension status (1 for yes, 0 for no).}
}
\value{
A numeric vector of eGDR values.
}
\description{
This function calculates the eGDR using waist circumference, hypertension status, and HbA1c values.
It works element-wise for vectors of HbA1c, waist, and hypertension.
}
\note{
\subsection{eGDR: Estimated Glucose Disposal Rate:}{

\if{html}{\out{<div class="sourceCode">}}\preformatted{# eGDR is calculated using the following formula:
eGDR = 21.158 + (-0.09 * waist circumference [cm]) + (−3.407 * hypertension [yes=1/no=0]) + (−0.551 * HbA1c (\%))
}\if{html}{\out{</div>}}

In UKB, HbA1c is measured in \strong{mmol/mol (IFCC)}, and needs to be converted to \strong{\% (NGSP)}.
\itemize{
\item \strong{HbA1c (mmol/mol)} (Field ID: 30750)
\itemize{
\item Conversion formula: NGSP (\%) = \link{0.09148 * IFCC} + 2.152 \href{https://ngsp.org/ifcc.asp}{Ref}
}
\item \strong{Waist circumference} (Field ID: 48)
\item \strong{hypertension} need to extract from icd10 or first occurrence records first.
}
}
}
\references{
\enumerate{
\item \url{https://diabetesjournals.org/care/article/36/8/2280/32950/Use-of-the-Estimated-Glucose-Disposal-Rate-as-a} (Diabetes Care, 2013 July)
\item \url{https://pmc.ncbi.nlm.nih.gov/articles/PMC11439291/#Sec2} (Cardiovasc Diabetol, 2024 Sep)
}
}
