% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{leo_ethnicity}
\alias{leo_ethnicity}
\title{Recode ethnicity to White vs non-White
https://biobank.ndph.ox.ac.uk/showcase/field.cgi?id=21000}
\usage{
leo_ethnicity(ethnicity, ...)
}
\arguments{
\item{ethnicity}{Numeric/character vector, usually from p21000.}
}
\value{
Factor with levels "White", "Non-White";
NA for -1/-3/missing.
}
\description{
Recode ethnicity to White vs non-White
https://biobank.ndph.ox.ac.uk/showcase/field.cgi?id=21000
}
