% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{leo_gender}
\alias{leo_gender}
\title{Recode gender (1 = Male, 0 = Female)
https://biobank.ndph.ox.ac.uk/showcase/field.cgi?id=31}
\usage{
leo_gender(gender, ...)
}
\arguments{
\item{gender}{Numeric/character vector, usually from p31.}
}
\value{
Integer vector (1 = Male, 0 = Female, NA for negative/special codes).
}
\description{
Recode gender (1 = Male, 0 = Female)
https://biobank.ndph.ox.ac.uk/showcase/field.cgi?id=31
}
