% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_vars.R
\name{leo_smoking_status}
\alias{leo_smoking_status}
\title{Recode smoking status (Never/Previous/Current)
https://biobank.ndph.ox.ac.uk/showcase/field.cgi?id=20116}
\usage{
leo_smoking_status(smoking_status, ...)
}
\arguments{
\item{smoking_status}{Numeric/character vector, usually from p20116.}
}
\value{
Factor with levels "Never", "Previous", "Current";
NA for negative/special codes or others.
}
\description{
Recode smoking status (Never/Previous/Current)
https://biobank.ndph.ox.ac.uk/showcase/field.cgi?id=20116
}
