% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype.R
\name{ukb_hla_header}
\alias{ukb_hla_header}
\title{ukb_hla_header}
\usage{
ukb_hla_header()
}
\value{
A data.frame with one column \code{name}.
}
\description{
Read UKB HLA v2 header line.
}
\details{
This function streams the UK Biobank HLA v2 header file, which is
described at: https://biobank.ctsu.ox.ac.uk/crystal/refer.cgi?id=2182
}
\examples{
ukb_hla_header <- ukb_hla_header()
head(ukb_hla_header)
}
